backend/requirement.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
requests
sentence-transformers
langchain-community
langchain-voyageai
langchain
langchain_openai
chromadb
python-dotenv

PILLOW
requests
customtkinter
tk
fitz